+++ src/raycasting.c
@@ -6,7 +6,7 @@
 /*   By: gbonnard <gbonnard@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/14 18:33:58 by gbonnard          #+#    #+#             */
-/*   Updated: 2023/11/30 19:37:52 by gbonnard         ###   ########.fr       */
+/*   Updated: 2023/12/02 12:51:34 by roman            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -69,7 +69,7 @@ int	raycasting(t_data *data)
 
 	while (data->ray.x < data->screen_width)
 	{
-		data->ray.camerax = 2 * data->ray.x / (double)data->screen_width - 1;
+		data->ray.camerax = 2 * data->ray.x / (float)data->screen_width - 1;
 		data->ray.raydirx = data->ray.dirx
 			+ data->ray.planex * data->ray.camerax;
 		data->ray.raydiry = data->ray.diry
